+++ lib/src/widgets/widgets.dart
@@ -1,1 +1,3 @@
 export 'value_watch.dart';
+export 'watch.dart';
+export 'watch_all.dart';