+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+### 2.0.0
+
+#### New Features
+
+- **Stream to Watcher**: Introduced `toWatcher` extension for streams, enabling conversion of any stream into
+  a `Watcher`.
+
+- **Cached Data Access**: Added `cache` getter in `CachedWatcher` for direct cached data access outside the `read`
+  function.
+
+- **isDisposed in Watcher**: Added `isDisposed` in the Watcher. it is useful when you want to safely do some operations
+  on the `Watcher` without getting exception if it gets disposed.
+
+- **Watch Widget**: New `Watch` widget added for reacting to changes in any `Listenable`, not limited to `Watcher`.
+
+- **WatchAll Widget**: Introduced `WatchAll` widget, similar to `Watch` but for a list of `Listenable` instances,
+  monitoring changes in each.
+
+#### Breaking Changes
+
+- **Widget Renaming**: Renamed `ValueWatch` Widget to `WatchValue`.
+
+- **Extension Modification**: Modified `.watch` extension; it no longer provides the value in the builder function.
+  Use `.watchValue` instead.
+
+- **CachedWatcher Update**: `CachedWatcher` is now abstract, requiring subclasses to implement `read` and `write`
+  methods for enhanced custom data handling.
+
+  **Changes**:
+  - Before: `final typeCachedWatcher = Type().cachedWatcher(read: .., write: ..);`
+  - Now: `final typeCachedWatcher = TypeCachedWatcher(); // TypeCachedWatcher extends CachedWatcher`
+  - Primitive types usage remains the same: `final counter = 0.cachedWatcher('counter_key');`
+
 ### 1.0.4
 
 - This version returns the `ListValueNotifierEx` Accidentally removed from the previous version.
@@ -15,11 +48,14 @@
 ### Example Usage
 
 ```dart
+
 final boolWatcher = true.watcher;
-boolWatcher.toggle();
+boolWatcher.toggle
+();
 
 final counter = 10.watcher;
-counter.increment(2);
+counter.increment
+(2);
 
 final urlWatcher = Uri.parse("https://example.com").watcher;
 urlWatcher.updatePath("/newpage"); // without the extension u would use `urlWatcher.value.updatePath("/newpage");`